Refactor rollers to belongto serializer subsystem package
