+++ src/main/java/frc/robot/subsystems/serializer/Serializer.java
@@ -1,0 +1,5 @@
+package frc.robot.subsystems.serializer;
+
+public class Serializer {
+    
+}
