+++ src/main/java/frc/robot/subsystems/serializer/sensors/SensorHal.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.serializer;
+package frc.robot.subsystems.serializer.sensors;
 
 import edu.wpi.first.wpilibj.DigitalInput;
