+++ src/main/java/frc/robot/subsystems/serializer/sensors/SensorHalI.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.serializer;
+package frc.robot.subsystems.serializer.sensors;
 
 public interface SensorHalI {
     public boolean getFrontBreak();
